implemented most of command handler

+++ Langer Yoshi/test worlds/controllers/Main_Controller/src/Command.h
@@ -2,8 +2,8 @@
 #define COMMAND_H
 
 struct Command {
-	int rotation;
-	float distance;
+	double rotation;
+	double distance;
 };
